🩹 [Patch]: Update linter and git configurations and license year (#7)
## Description

This pull request includes several changes to configuration files and a
minor update to the license file. The most important changes include
updates to linter configurations and the license year.

Linter configuration updates:

*
[`.github/linters/.jscpd.json`](diffhunk://#diff-557094e283c00b23265c1c75872f41c6b1a524a00f0d99dd68ebd22cb63bfdd6R1-R10):
Added a new configuration file for the `jscpd` linter to set the
threshold to 0, specify reporters, and ignore test files.
*
[`.github/linters/.powershell-psscriptanalyzer.psd1`](diffhunk://#diff-aae69c9d6774628ed181eacf53aee0f38eb6c2f53492cf3a5b7f7bdb6ef43b6aL1-L16):
Updated the configuration for `PSScriptAnalyzer` to enable and configure
specific rules, such as `PSAlignAssignmentStatement`,
`PSAvoidLongLines`, and `PSUseConsistentIndentation`.
*
[`.github/workflows/Linter.yml`](diffhunk://#diff-482e65806ed9e4a7320f14964764086b91fed4a28d12e4efde1776472e147e79R30):
Added the `VALIDATE_JSON_PRETTIER` environment variable and set it to
false to disable JSON validation with Prettier.

License update:

*
[`LICENSE`](diffhunk://#diff-c693279643b8cd5d248172d9c22cb7cf4ed163a3c98c8a3f69c2717edd3eacb7L3-R3):
Updated the copyright year from 2024 to 2025.

## Type of change

<!-- Use the check-boxes [x] on the options that are relevant. -->

- [ ] 📖 [Docs]
- [ ] 🪲 [Fix]
- [x] 🩹 [Patch]
- [ ] ⚠️ [Security fix]
- [ ] 🚀 [Feature]
- [ ] 🌟 [Breaking change]

## Checklist

<!-- Use the check-boxes [x] on the options that are relevant. -->

- [x] I have performed a self-review of my own code
- [x] I have commented my code, particularly in hard-to-understand areas